+++ cli/src/lib.rs
@@ -1,5 +1,6 @@
 use crate::config::{
-    AnchorPackage, Config, ConfigOverride, Manifest, ProgramDeployment, ProgramWorkspace, WithPath,
+    AnchorPackage, Config, ConfigOverride, Manifest, ProgramDeployment, ProgramWorkspace, Test,
+    WithPath,
 };
 use anchor_client::Cluster;
 use anchor_lang::idl::{IdlAccount, IdlInstruction};
@@ -199,6 +200,25 @@ pub enum Command {
         #[clap(subcommand)]
         subcmd: KeysCommand,
     },
+    /// Localnet commands.
+    Localnet {
+        /// Flag to skip building the program in the workspace,
+        /// use this to save time when running test and the program code is not altered.
+        #[clap(long)]
+        skip_build: bool,
+        /// Use this flag if you want to run tests against previously deployed
+        /// programs.
+        #[clap(long)]
+        skip_deploy: bool,
+        /// Arguments to pass to the underlying `cargo build-bpf` command.
+        #[clap(
+            required = false,
+            takes_value = true,
+            multiple_values = true,
+            last = true
+        )]
+        cargo_args: Vec<String>,
+    },
 }
 
 #[derive(Debug, Clap)]
@@ -350,6 +370,11 @@ pub fn entry(opts: Opts) -> Result<()> {
             cargo_args,
         } => publish(&opts.cfg_override, program, cargo_args),
         Command::Keys { subcmd } => keys(&opts.cfg_override, subcmd),
+        Command::Localnet {
+            skip_build,
+            skip_deploy,
+            cargo_args,
+        } => localnet(&opts.cfg_override, skip_build, skip_deploy, cargo_args),
     }
 }
 
@@ -496,19 +521,18 @@ pub fn build(
     }
 
     let cfg = Config::discover(cfg_override)?.expect("Not in workspace.");
+    let cfg_parent = cfg.path().parent().expect("Invalid Anchor.toml");
+
     let cargo = Manifest::discover()?;
+
+    fs::create_dir_all(cfg_parent.join("target/idl"))?;
+    fs::create_dir_all(cfg_parent.join("target/types"))?;
 
     let idl_out = match idl {
         Some(idl) => Some(PathBuf::from(idl)),
-        None => {
-            let cfg_parent = match cfg.path().parent() {
-                None => return Err(anyhow!("Invalid Anchor.toml")),
-                Some(parent) => parent,
-            };
-            fs::create_dir_all(cfg_parent.join("target/idl"))?;
-            Some(cfg_parent.join("target/idl"))
-        }
+        None => Some(cfg_parent.join("target/idl")),
     };
+    let idl_ts_out = Some(cfg_parent.join("target/types"));
 
     let solana_version = match solana_version.is_some() {
         true => solana_version,
@@ -521,6 +545,7 @@ pub fn build(
             &cfg,
             cfg.path(),
             idl_out,
+            idl_ts_out,
             verifiable,
             solana_version,
             stdout,
@@ -532,6 +557,7 @@ pub fn build(
             &cfg,
             cfg.path(),
             idl_out,
+            idl_ts_out,
             verifiable,
             solana_version,
             stdout,
@@ -543,6 +569,7 @@ pub fn build(
             &cfg,
             cargo.path().to_path_buf(),
             idl_out,
+            idl_ts_out,
             verifiable,
             solana_version,
             stdout,
@@ -561,6 +588,7 @@ fn build_all(
     cfg: &WithPath<Config>,
     cfg_path: &Path,
     idl_out: Option<PathBuf>,
+    idl_ts_out: Option<PathBuf>,
     verifiable: bool,
     solana_version: Option<String>,
     stdout: Option<File>, // Used for the package registry server.
@@ -576,6 +604,7 @@ fn build_all(
                     cfg,
                     p.join("Cargo.toml"),
                     idl_out.clone(),
+                    idl_ts_out.clone(),
                     verifiable,
                     solana_version.clone(),
                     stdout.as_ref().map(|f| f.try_clone()).transpose()?,
@@ -596,6 +625,7 @@ fn build_cwd(
     cfg: &WithPath<Config>,
     cargo_toml: PathBuf,
     idl_out: Option<PathBuf>,
+    idl_ts_out: Option<PathBuf>,
     verifiable: bool,
     solana_version: Option<String>,
     stdout: Option<File>,
@@ -607,7 +637,7 @@ fn build_cwd(
         Some(p) => std::env::set_current_dir(&p)?,
     };
     match verifiable {
-        false => _build_cwd(idl_out, cargo_args),
+        false => _build_cwd(idl_out, idl_ts_out, cargo_args),
         true => build_cwd_verifiable(cfg, cargo_toml, solana_version, stdout, stderr),
     }
 }
@@ -625,6 +655,7 @@ fn build_cwd_verifiable(
     let workspace_dir = cfg.path().parent().unwrap().canonicalize()?;
     fs::create_dir_all(workspace_dir.join("target/verifiable"))?;
     fs::create_dir_all(workspace_dir.join("target/idl"))?;
+    fs::create_dir_all(workspace_dir.join("target/types"))?;
 
     let container_name = "anchor-program";
 
@@ -670,11 +701,19 @@ fn build_cwd_verifiable(
     }
 
     // Build the idl.
+    println!("Extracting the IDL");
     if let Ok(Some(idl)) = extract_idl("src/lib.rs") {
-        println!("Extracting the IDL");
+        // Write out the JSON file.
+        println!("Writing the IDL file");
         let out_file = workspace_dir.join(format!("target/idl/{}.json", idl.name));
         write_idl(&idl, OutFile::File(out_file))?;
+
+        // Write out the TypeScript type.
+        println!("Writing the .ts file");
+        let ts_file = workspace_dir.join(format!("target/types/{}.ts", idl.name));
+        fs::write(&ts_file, template::idl_ts(&idl)?)?;
     }
+    println!("Build success");
 
     result
 }
@@ -836,7 +875,11 @@ fn docker_build(
     Ok(())
 }
 
-fn _build_cwd(idl_out: Option<PathBuf>, cargo_args: Vec<String>) -> Result<()> {
+fn _build_cwd(
+    idl_out: Option<PathBuf>,
+    idl_ts_out: Option<PathBuf>,
+    cargo_args: Vec<String>,
+) -> Result<()> {
     let exit = std::process::Command::new("cargo")
         .arg("build-bpf")
         .args(cargo_args)
@@ -850,12 +893,22 @@ fn _build_cwd(idl_out: Option<PathBuf>, cargo_args: Vec<String>) -> Result<()> {
 
     // Always assume idl is located ar src/lib.rs.
     if let Some(idl) = extract_idl("src/lib.rs")? {
+        // JSON out path.
         let out = match idl_out {
             None => PathBuf::from(".").join(&idl.name).with_extension("json"),
             Some(o) => PathBuf::from(&o.join(&idl.name).with_extension("json")),
         };
+        // TS out path.
+        let ts_out = match idl_ts_out {
+            None => PathBuf::from(".").join(&idl.name).with_extension("ts"),
+            Some(o) => PathBuf::from(&o.join(&idl.name).with_extension("ts")),
+        };
 
+        // Write out the JSON file.
         write_idl(&idl, OutFile::File(out))?;
+
+        // Write out the TypeScript type.
+        fs::write(ts_out, template::idl_ts(&idl)?)?;
     }
 
     Ok(())
@@ -903,7 +956,8 @@ fn verify(
         .join("target/verifiable/")
         .join(format!("{}.so", binary_name));
 
-    let bin_ver = verify_bin(program_id, &bin_path, cfg.provider.cluster.url())?;
+    let url = cluster_url(&cfg);
+    let bin_ver = verify_bin(program_id, &bin_path, &url)?;
     if !bin_ver.is_verified {
         println!("Error: Binaries don't match");
         std::process::exit(1);
@@ -1028,7 +1082,8 @@ pub enum BinVerificationState {
 // Fetches an IDL for the given program_id.
 fn fetch_idl(cfg_override: &ConfigOverride, idl_addr: Pubkey) -> Result<Idl> {
     let cfg = Config::discover(cfg_override)?.expect("Inside a workspace");
-    let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+    let url = cluster_url(&cfg);
+    let client = RpcClient::new(url);
 
     let mut account = client
         .get_account_with_commitment(&idl_addr, CommitmentConfig::processed())?
@@ -1091,7 +1146,7 @@ fn idl_init(cfg_override: &ConfigOverride, program_id: Pubkey, idl_filepath: Str
     with_workspace(cfg_override, |cfg| {
         let keypair = cfg.provider.wallet.to_string();
 
-        let bytes = std::fs::read(idl_filepath)?;
+        let bytes = fs::read(idl_filepath)?;
         let idl: Idl = serde_json::from_reader(&*bytes)?;
 
         let idl_address = create_idl_account(cfg, &keypair, &program_id, &idl)?;
@@ -1109,7 +1164,7 @@ fn idl_write_buffer(
     with_workspace(cfg_override, |cfg| {
         let keypair = cfg.provider.wallet.to_string();
 
-        let bytes = std::fs::read(idl_filepath)?;
+        let bytes = fs::read(idl_filepath)?;
         let idl: Idl = serde_json::from_reader(&*bytes)?;
 
         let idl_buffer = create_idl_buffer(cfg, &keypair, &program_id, &idl)?;
@@ -1125,7 +1180,8 @@ fn idl_set_buffer(cfg_override: &ConfigOverride, program_id: Pubkey, buffer: Pub
     with_workspace(cfg_override, |cfg| {
         let keypair = solana_sdk::signature::read_keypair_file(&cfg.provider.wallet.to_string())
             .map_err(|_| anyhow!("Unable to read keypair file"))?;
-        let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+        let url = cluster_url(cfg);
+        let client = RpcClient::new(url);
 
         // Instruction to set the buffer onto the IdlAccount.
         let set_buffer_ix = {
@@ -1177,7 +1233,8 @@ fn idl_upgrade(
 
 fn idl_authority(cfg_override: &ConfigOverride, program_id: Pubkey) -> Result<()> {
     with_workspace(cfg_override, |cfg| {
-        let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+        let url = cluster_url(cfg);
+        let client = RpcClient::new(url);
         let idl_address = {
             let account = client
                 .get_account_with_commitment(&program_id, CommitmentConfig::processed())?
@@ -1214,7 +1271,8 @@ fn idl_set_authority(
         };
         let keypair = solana_sdk::signature::read_keypair_file(&cfg.provider.wallet.to_string())
             .map_err(|_| anyhow!("Unable to read keypair file"))?;
-        let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+        let url = cluster_url(cfg);
+        let client = RpcClient::new(url);
 
         // Instruction data.
         let data =
@@ -1284,7 +1342,8 @@ fn idl_write(cfg: &Config, program_id: &Pubkey, idl: &Idl, idl_address: Pubkey) 
     // Misc.
     let keypair = solana_sdk::signature::read_keypair_file(&cfg.provider.wallet.to_string())
         .map_err(|_| anyhow!("Unable to read keypair file"))?;
-    let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+    let url = cluster_url(cfg);
+    let client = RpcClient::new(url);
 
     // Serialize and compress the idl.
     let idl_data = {
@@ -1359,8 +1418,9 @@ fn write_idl(idl: &Idl, out: OutFile) -> Result<()> {
     let idl_json = serde_json::to_string_pretty(idl)?;
     match out {
         OutFile::Stdout => println!("{}", idl_json),
-        OutFile::File(out) => std::fs::write(out, idl_json)?,
+        OutFile::File(out) => fs::write(out, idl_json)?,
     };
+
     Ok(())
 }
 
@@ -1409,13 +1469,15 @@ fn test(
         if is_localnet && (!skip_local_validator) {
             let flags = match skip_deploy {
                 true => None,
-                false => Some(genesis_flags(cfg)?),
+                false => Some(validator_flags(cfg)?),
             };
-            validator_handle = Some(start_test_validator(cfg, flags)?);
+            validator_handle = Some(start_test_validator(cfg, flags, true)?);
         }
 
+        let url = cluster_url(cfg);
+
         // Setup log reader.
-        let log_streams = stream_logs(cfg);
+        let log_streams = stream_logs(cfg, &url);
 
         // Run the tests.
         let test_result: Result<_> = {
@@ -1432,7 +1494,7 @@ fn test(
 
             std::process::Command::new(program)
                 .args(args)
-                .env("ANCHOR_PROVIDER_URL", cfg.provider.cluster.url())
+                .env("ANCHOR_PROVIDER_URL", url)
                 .env("ANCHOR_WALLET", cfg.provider.wallet.to_string())
                 .stdout(Stdio::inherit())
                 .stderr(Stdio::inherit())
@@ -1475,9 +1537,10 @@ fn test(
     })
 }
 
-// Returns the solana-test-validator flags to embed the workspace programs
-// in the genesis block. This allows us to run tests without every deploying.
-fn genesis_flags(cfg: &WithPath<Config>) -> Result<Vec<String>> {
+// Returns the solana-test-validator flags. This will embed the workspace
+// programs in the genesis block so we don't have to deploy every time. It also
+// allows control of other solana-test-validator features.
+fn validator_flags(cfg: &WithPath<Config>) -> Result<Vec<String>> {
     let programs = cfg.programs.get(&Cluster::Localnet);
 
     let mut flags = Vec::new();
@@ -1506,20 +1569,50 @@ fn genesis_flags(cfg: &WithPath<Config>) -> Result<Vec<String>> {
             write_idl(idl, OutFile::File(idl_out))?;
         }
     }
+
     if let Some(test) = cfg.test.as_ref() {
-        for entry in &test.genesis {
-            flags.push("--bpf-program".to_string());
-            flags.push(entry.address.clone());
-            flags.push(entry.program.clone());
+        if let Some(genesis) = &test.genesis {
+            for entry in genesis {
+                let program_path = Path::new(&entry.program);
+                if !program_path.exists() {
+                    return Err(anyhow!(
+                        "Program in genesis configuration does not exist at path: {}",
+                        program_path.display()
+                    ));
+                }
+                flags.push("--bpf-program".to_string());
+                flags.push(entry.address.clone());
+                flags.push(entry.program.clone());
+            }
+        }
+        if let Some(clone) = &test.clone {
+            for entry in clone {
+                flags.push("--clone".to_string());
+                flags.push(entry.address.clone());
+            }
+        }
+        if let Some(validator) = &test.validator {
+            for (key, value) in serde_json::to_value(validator)?.as_object().unwrap() {
+                if key == "ledger" {
+                    continue;
+                };
+                flags.push(format!("--{}", key.replace("_", "-")));
+                if let serde_json::Value::String(v) = value {
+                    flags.push(v.to_string());
+                } else {
+                    flags.push(value.to_string());
+                }
+            }
         }
     }
+
     Ok(flags)
 }
 
-fn stream_logs(config: &WithPath<Config>) -> Result<Vec<std::process::Child>> {
+fn stream_logs(config: &WithPath<Config>, rpc_url: &str) -> Result<Vec<std::process::Child>> {
     let program_logs_dir = ".anchor/program-logs";
     if Path::new(program_logs_dir).exists() {
-        std::fs::remove_dir_all(program_logs_dir)?;
+        fs::remove_dir_all(program_logs_dir)?;
     }
     fs::create_dir_all(program_logs_dir)?;
     let mut handles = vec![];
@@ -1542,23 +1635,25 @@ fn stream_logs(config: &WithPath<Config>) -> Result<Vec<std::process::Child>> {
             .arg("logs")
             .arg(metadata.address)
             .arg("--url")
-            .arg(config.provider.cluster.url())
+            .arg(rpc_url)
             .stdout(stdio)
             .spawn()?;
         handles.push(child);
     }
     if let Some(test) = config.test.as_ref() {
-        for entry in &test.genesis {
-            let log_file = File::create(format!("{}/{}.log", program_logs_dir, entry.address))?;
-            let stdio = std::process::Stdio::from(log_file);
-            let child = std::process::Command::new("solana")
-                .arg("logs")
-                .arg(entry.address.clone())
-                .arg("--url")
-                .arg(config.provider.cluster.url())
-                .stdout(stdio)
-                .spawn()?;
-            handles.push(child);
+        if let Some(genesis) = &test.genesis {
+            for entry in genesis {
+                let log_file = File::create(format!("{}/{}.log", program_logs_dir, entry.address))?;
+                let stdio = std::process::Stdio::from(log_file);
+                let child = std::process::Command::new("solana")
+                    .arg("logs")
+                    .arg(entry.address.clone())
+                    .arg("--url")
+                    .arg(rpc_url)
+                    .stdout(stdio)
+                    .spawn()?;
+                handles.push(child);
+            }
         }
     }
 
@@ -1570,36 +1665,48 @@ pub struct IdlTestMetadata {
     address: String,
 }
 
-fn start_test_validator(cfg: &Config, flags: Option<Vec<String>>) -> Result<Child> {
-    fs::create_dir_all(".anchor")?;
-    let test_ledger_filename = ".anchor/test-ledger";
-    let test_ledger_log_filename = ".anchor/test-ledger-log.txt";
-
-    if Path::new(test_ledger_filename).exists() {
-        std::fs::remove_dir_all(test_ledger_filename)?;
-    }
-    if Path::new(test_ledger_log_filename).exists() {
-        std::fs::remove_file(test_ledger_log_filename)?;
-    }
+fn start_test_validator(
+    cfg: &Config,
+    flags: Option<Vec<String>>,
+    test_log_stdout: bool,
+) -> Result<Child> {
+    //
+    let (test_ledger_directory, test_ledger_log_filename) = test_validator_file_paths(cfg);
 
     // Start a validator for testing.
-    let test_validator_stdout = File::create(test_ledger_log_filename)?;
-    let test_validator_stderr = test_validator_stdout.try_clone()?;
-    let validator_handle = std::process::Command::new("solana-test-validator")
+    let (test_validator_stdout, test_validator_stderr) = match test_log_stdout {
+        true => {
+            let test_validator_stdout_file = File::create(&test_ledger_log_filename)?;
+            let test_validator_sterr_file = test_validator_stdout_file.try_clone()?;
+            (
+                Stdio::from(test_validator_stdout_file),
+                Stdio::from(test_validator_sterr_file),
+            )
+        }
+        false => (Stdio::inherit(), Stdio::inherit()),
+    };
+
+    let rpc_url = test_validator_rpc_url(cfg);
+
+    let mut validator_handle = std::process::Command::new("solana-test-validator")
         .arg("--ledger")
-        .arg(test_ledger_filename)
+        .arg(test_ledger_directory)
         .arg("--mint")
         .arg(cfg.wallet_kp()?.pubkey().to_string())
         .args(flags.unwrap_or_default())
-        .stdout(Stdio::from(test_validator_stdout))
-        .stderr(Stdio::from(test_validator_stderr))
+        .stdout(test_validator_stdout)
+        .stderr(test_validator_stderr)
         .spawn()
         .map_err(|e| anyhow::format_err!("{}", e.to_string()))?;
 
     // Wait for the validator to be ready.
-    let client = RpcClient::new("http://localhost:8899".to_string());
+    let client = RpcClient::new(rpc_url);
     let mut count = 0;
-    let ms_wait = 5000;
+    let ms_wait = cfg
+        .test
+        .as_ref()
+        .and_then(|test| test.startup_wait)
+        .unwrap_or(5_000);
     while count < ms_wait {
         let r = client.get_recent_blockhash();
         if r.is_ok() {
@@ -1608,17 +1715,71 @@ fn start_test_validator(cfg: &Config, flags: Option<Vec<String>>) -> Result<Chil
         std::thread::sleep(std::time::Duration::from_millis(1));
         count += 1;
     }
-    if count == 5000 {
-        println!("Unable to start test validator.");
+    if count == ms_wait {
+        eprintln!(
+            "Unable to start test validator. Check {} for errors.",
+            test_ledger_log_filename
+        );
+        validator_handle.kill()?;
         std::process::exit(1);
     }
-
     Ok(validator_handle)
+}
+
+// Return the URL that solana-test-validator should be running on given the
+// configuration
+fn test_validator_rpc_url(cfg: &Config) -> String {
+    match &cfg.test.as_ref() {
+        Some(Test {
+            validator: Some(validator),
+            ..
+        }) => format!("http://{}:{}", validator.bind_address, validator.rpc_port),
+        _ => "http://localhost:8899".to_string(),
+    }
+}
+
+// Setup and return paths to the solana-test-validator ledger directory and log
+// files given the configuration
+fn test_validator_file_paths(cfg: &Config) -> (String, String) {
+    let ledger_directory = match &cfg.test.as_ref() {
+        Some(Test {
+            validator: Some(validator),
+            ..
+        }) => &validator.ledger,
+        _ => ".anchor/test-ledger",
+    };
+
+    if !Path::new(&ledger_directory).is_relative() {
+        // Prevent absolute paths to avoid someone using / or similar, as the
+        // directory gets removed
+        eprintln!("Ledger directory {} must be relative", ledger_directory);
+        std::process::exit(1);
+    }
+    if Path::new(&ledger_directory).exists() {
+        fs::remove_dir_all(&ledger_directory).unwrap();
+    }
+
+    fs::create_dir_all(&ledger_directory).unwrap();
+
+    (
+        ledger_directory.to_string(),
+        format!("{}/test-ledger-log.txt", ledger_directory),
+    )
+}
+
+fn cluster_url(cfg: &Config) -> String {
+    let is_localnet = cfg.provider.cluster == Cluster::Localnet;
+    match is_localnet {
+        // Cluster is Localnet, assume the intent is to use the configuration
+        // for solana-test-validator
+        true => test_validator_rpc_url(cfg),
+        false => cfg.provider.cluster.url().to_string(),
+    }
 }
 
 fn deploy(cfg_override: &ConfigOverride, program_str: Option<String>) -> Result<()> {
     with_workspace(cfg_override, |cfg| {
-        let url = cfg.provider.cluster.url().to_string();
+        let url = cluster_url(cfg);
         let keypair = cfg.provider.wallet.to_string();
 
         // Deploy the programs.
@@ -1692,11 +1853,12 @@ fn upgrade(
     let program_filepath = path.canonicalize()?.display().to_string();
 
     with_workspace(cfg_override, |cfg| {
+        let url = cluster_url(cfg);
         let exit = std::process::Command::new("solana")
             .arg("program")
             .arg("deploy")
             .arg("--url")
-            .arg(cfg.provider.cluster.url())
+            .arg(url)
             .arg("--keypair")
             .arg(&cfg.provider.wallet.to_string())
             .arg("--program-id")
@@ -1724,7 +1886,8 @@ fn create_idl_account(
     let idl_address = IdlAccount::address(program_id);
     let keypair = solana_sdk::signature::read_keypair_file(keypair_path)
         .map_err(|_| anyhow!("Unable to read keypair file"))?;
-    let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+    let url = cluster_url(cfg);
+    let client = RpcClient::new(url);
     let idl_data = serialize_idl(idl)?;
 
     // Run `Create instruction.
@@ -1777,7 +1940,8 @@ fn create_idl_buffer(
 ) -> Result<Pubkey> {
     let keypair = solana_sdk::signature::read_keypair_file(keypair_path)
         .map_err(|_| anyhow!("Unable to read keypair file"))?;
-    let client = RpcClient::new(cfg.provider.cluster.url().to_string());
+    let url = cluster_url(cfg);
+    let client = RpcClient::new(url);
 
     let buffer = Keypair::generate(&mut OsRng);
 
@@ -1850,7 +2014,7 @@ fn migrate(cfg_override: &ConfigOverride) -> Result<()> {
     with_workspace(cfg_override, |cfg| {
         println!("Running migration deploy script");
 
-        let url = cfg.provider.cluster.url().to_string();
+        let url = cluster_url(cfg);
         let cur_dir = std::env::current_dir()?;
 
         let use_ts =
@@ -1865,7 +2029,7 @@ fn migrate(cfg_override: &ConfigOverride) -> Result<()> {
             let module_path = cur_dir.join("migrations/deploy.ts");
             let deploy_script_host_str =
                 template::deploy_ts_script_host(&url, &module_path.display().to_string());
-            std::fs::write("deploy.ts", deploy_script_host_str)?;
+            fs::write("deploy.ts", deploy_script_host_str)?;
             std::process::Command::new("ts-node")
                 .arg("deploy.ts")
                 .env("ANCHOR_WALLET", cfg.provider.wallet.to_string())
@@ -1876,7 +2040,7 @@ fn migrate(cfg_override: &ConfigOverride) -> Result<()> {
             let module_path = cur_dir.join("migrations/deploy.js");
             let deploy_script_host_str =
                 template::deploy_js_script_host(&url, &module_path.display().to_string());
-            std::fs::write("deploy.js", deploy_script_host_str)?;
+            fs::write("deploy.js", deploy_script_host_str)?;
             std::process::Command::new("node")
                 .arg("deploy.js")
                 .env("ANCHOR_WALLET", cfg.provider.wallet.to_string())
@@ -1897,8 +2061,8 @@ fn migrate(cfg_override: &ConfigOverride) -> Result<()> {
 
 fn set_workspace_dir_or_exit() {
     let d = match Config::discover(&ConfigOverride::default()) {
-        Err(_) => {
-            println!("Not in anchor workspace.");
+        Err(err) => {
+            println!("Workspace configuration error: {}", err);
             std::process::exit(1);
         }
         Ok(d) => d,
@@ -1978,7 +2142,7 @@ fn shell(cfg_override: &ConfigOverride) -> Result<()> {
                     .map(|(name, pd)| {
                         if let Some(idl_fp) = &pd.idl {
                             let file_str =
-                                std::fs::read_to_string(idl_fp).expect("Unable to read IDL file");
+                                fs::read_to_string(idl_fp).expect("Unable to read IDL file");
                             let idl = serde_json::from_str(&file_str).expect("Idl not readable");
                             idls.insert(name.clone(), idl);
                         }
@@ -2004,11 +2168,8 @@ fn shell(cfg_override: &ConfigOverride) -> Result<()> {
                     .collect::<Vec<ProgramWorkspace>>(),
             }
         };
-        let js_code = template::node_shell(
-            cfg.provider.cluster.url(),
-            &cfg.provider.wallet.to_string(),
-            programs,
-        )?;
+        let url = cluster_url(cfg);
+        let js_code = template::node_shell(&url, &cfg.provider.wallet.to_string(), programs)?;
         let mut child = std::process::Command::new("node")
             .args(&["-e", &js_code, "-i", "--experimental-repl-await"])
             .stdout(Stdio::inherit())
@@ -2026,6 +2187,7 @@ fn shell(cfg_override: &ConfigOverride) -> Result<()> {
 
 fn run(cfg_override: &ConfigOverride, script: String) -> Result<()> {
     with_workspace(cfg_override, |cfg| {
+        let url = cluster_url(cfg);
         let script = cfg
             .scripts
             .get(&script)
@@ -2033,7 +2195,7 @@ fn run(cfg_override: &ConfigOverride, script: String) -> Result<()> {
         let exit = std::process::Command::new("bash")
             .arg("-c")
             .arg(&script)
-            .env("ANCHOR_PROVIDER_URL", cfg.provider.cluster.url())
+            .env("ANCHOR_PROVIDER_URL", url)
             .env("ANCHOR_WALLET", cfg.provider.wallet.to_string())
             .stdout(Stdio::inherit())
             .stderr(Stdio::inherit())
@@ -2147,7 +2309,7 @@ fn publish(
             // Skip target dir.
             if !path_str.contains("target/") && !path_str.contains("/target") {
                 // Only add the file if it's not empty.
-                let metadata = std::fs::File::open(&e)?.metadata()?;
+                let metadata = fs::File::open(&e)?.metadata()?;
                 if metadata.len() > 0 {
                     println!("PACKING: {}", e.display().to_string());
                     if e.is_dir() {
@@ -2258,6 +2420,59 @@ fn keys_list(cfg_override: &ConfigOverride) -> Result<()> {
         println!("{}: {}", program.lib_name, pubkey.to_string());
     }
     Ok(())
+}
+
+fn localnet(
+    cfg_override: &ConfigOverride,
+    skip_build: bool,
+    skip_deploy: bool,
+    cargo_args: Vec<String>,
+) -> Result<()> {
+    with_workspace(cfg_override, |cfg| {
+        // Build if needed.
+        if !skip_build {
+            build(
+                cfg_override,
+                None,
+                false,
+                None,
+                None,
+                None,
+                None,
+                cargo_args,
+            )?;
+        }
+
+        let flags = match skip_deploy {
+            true => None,
+            false => Some(validator_flags(cfg)?),
+        };
+
+        let validator_handle = &mut start_test_validator(cfg, flags, false)?;
+
+        // Setup log reader.
+        let url = test_validator_rpc_url(cfg);
+        let log_streams = stream_logs(cfg, &url);
+
+        std::io::stdin().lock().lines().next().unwrap().unwrap();
+
+        // Check all errors and shut down.
+        if let Err(err) = validator_handle.kill() {
+            println!(
+                "Failed to kill subprocess {}: {}",
+                validator_handle.id(),
+                err
+            );
+        }
+
+        for mut child in log_streams? {
+            if let Err(err) = child.kill() {
+                println!("Failed to kill subprocess {}: {}", child.id(), err);
+            }
+        }
+
+        Ok(())
+    })
 }
 
 // with_workspace ensures the current working directory is always the top level